Return an ActiveMerchant::Billing::Response on fail
This is a payment method, if it returns a boolean then
spree/payment/processing.rb will blow up. We are seeing this happen when
captured store credits are refunded and the handle_void_response method
tries to check the response.success?

Although I have not added a test for the last code path, I think we can
refactor it away when we're done with our changes.